+++ sdk/azidentity/client_secret_credential_test.go
@@ -6,6 +6,7 @@ package azidentity
 import (
 	"context"
 	"errors"
+	"reflect"
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
@@ -55,8 +56,8 @@ func TestClientSecretCredential_InvalidSecretLive(t *testing.T) {
 		t.Fatalf("failed to construct credential: %v", err)
 	}
 	tk, err := cred.GetToken(context.Background(), policy.TokenRequestOptions{Scopes: []string{liveTestScope}})
-	if tk != nil {
-		t.Fatal("GetToken returned a token")
+	if !reflect.ValueOf(tk).IsZero() {
+		t.Fatal("expected a zero value AccessToken")
 	}
 	var e *AuthenticationFailedError
 	if !errors.As(err, &e) {